+++ lib/operately_web/api/mutations/add_reaction.ex
@@ -12,6 +12,7 @@ defmodule OperatelyWeb.Api.Mutations.AddReaction do
   }
   alias Operately.Goals.Update
   alias Operately.Messages.Message
+  alias Operately.Projects.Retrospective
   alias Operately.Operations.ReactionAdding
 
   inputs do
@@ -53,6 +54,7 @@ defmodule OperatelyWeb.Api.Mutations.AddReaction do
   defp fetch_parent(id, person, type, parent_type) do
     case type do
       :project_check_in -> Projects.get_check_in_with_access_level(id, person.id)
+      :project_retrospective -> Retrospective.get(person, id: id)
       :comment_thread -> Comments.get_thread_with_activity_and_access_level(id, person.id)
       :goal_update -> Update.get(person, id: id)
       :message -> Message.get(person, id: id)
@@ -63,6 +65,7 @@ defmodule OperatelyWeb.Api.Mutations.AddReaction do
   defp check_permissions(parent, type, parent_type) do
     case type do
       :project_check_in -> Projects.Permissions.check(parent.requester_access_level, :can_comment_on_check_in)
+      :project_retrospective -> Projects.Permissions.check(parent.request_info.access_level, :can_comment_on_retrospective)
       :comment_thread -> Activities.Permissions.check(parent.activity.requester_access_level, :can_comment_on_thread)
       :goal_update -> Goals.Permissions.check(parent.request_info.access_level, :can_comment_on_update)
       :message -> Groups.Permissions.check(parent.request_info.access_level, :can_comment_on_discussions)
