+++ lib/operately_web/api/serializers/project_retrospective.ex
@@ -7,6 +7,7 @@ defimpl OperatelyWeb.Api.Serializable, for: Operately.Projects.Retrospective do
       content: Jason.encode!(retrospective.content),
       closed_at: OperatelyWeb.Api.Serializer.serialize(retrospective.closed_at),
       permissions: OperatelyWeb.Api.Serializer.serialize(retrospective.permissions),
+      reactions: OperatelyWeb.Api.Serializer.serialize(retrospective.reactions),
     }
   end
 end
